Fix bug causing  SAVE_ERROR to transition back to Empty instead of Form when closed
Fix by implementing the prevState callback method of setting history state in the transition function guard clause if replace argument is true. Ensure tests pass.

Due to this change, pass in true for the second argument ‘replace’ in the cancel appointment function when transitioning to the PROCESSING view so that when the ERROR_DELETE view is closed, it redirects to the SHOW view and not the CONFIRM view.

setHistory in the back function works fine but it could be changed to use the callback prevState method similar to the transition function made in this commit.

It seems the issue was that the new mode was not present in the history array when the if statement logic is executed while transition is being invoked. Meaning, the copied array in the previous implementation had the incorrect values.

Also, implement best practice by moving useEffect to the top of visualMode hook.